requirements.txt
streamlit
requests
pandas
pymysql
sshtunnel
python-dotenv